update meta descriptions, keywords, and index
A lot of the meta content was the same which isn't exactly good for SEO
or searching in windows explorer. The descriptions are mostly based on
the exercise info. As for the keywords, I've added some which are
relevant to the exercise, such as the book, page number, and problem
number. Searching a page number in windows explorer should bring up the
exercises for that specific page, likewise for problems (e.g. problem
V-A) and books (e.g. Genki I, Genki I Workbook, Genki II..).

I also updated the index to remove my personal monologue, and added some
additional information.