updated citation syntax
When citing pages from Genki, we'll be using the following syntax. (see
examples)

(VOLUME: PAGE; PROBLEM/SECTION)
Volume is optional when defined prior, such as in a heading that denotes
a section for that specific volume. Problem/Section is optional, but if
an exercise is based on a problem/section it MUST be included.

# Volume and page
- "You may want to read the Culture Note about Greetings and Bowing
(Genki I: p.37)"
- "Study Dialogue I (Genki I: p.84) and translate the conversation."

# Page and problem
- "Practice: Numbers (p.69; I-A)"
- "Workbook: Numbers (p.20; I, II, & III)"

# Page only (volume was defined prior)
- Volume definition: "Lesson 3: Making a Date (Genki I: p.84-101)"
- "Vocabulary: Entertainment and Sports (p.86)"
- "Culture Note: Japanese Houses (p.101)"