updated JS and heading attributes
- separated the Genki utilities for exercises from the main Javascript
file for all pages. These utilities are now located in Genki.js and are
only executed on exercises. This should improve performance on
non-exercise pages.
- updated Genki.local. It will now store a string based on the online
state: If offline, the value will be 'index.html'; if online the value
will be an empty string. You can use this property to append index.html
to links that are missed by the loop in javascript.js.
- added more ids to main headings, for anchors.
- added some CSS to correct the position of the anchor links in centered
elements.