fix readme quick links
Be me, updating the readme with quick links.
Think all is okay, so commits and pushes to master branch.
Decides to click on quick link and it leads to a 404 page.
My face when I forgot to include the hashes. (-_-;)
Don't be me; double and triple check your commits before pushing them.