add new exercises
- added the following exercises.

# Lesson 23 (3rd ed.)
- Reading Practice: 顔文字と絵文字
- Workbook: Kanji Writing Practice
- Workbook: Using Kanji

All exercises for the 3rd edition are now available. This is the last daily update. With this, the project is completed, and my work finished.

Thank you to all who have supported me along the way.. I would not have been able to make this far otherwise. So, once again, THANK YOU m(__)m